Implement CellStyle::update(..) and TableStyle::update(..).
Also set the context on the internal cell style to the document wide
StyleSet for cell styles. This was the last piece of the puzzle; cell
styles can now be created in the SM and set on table cells using
setStyle(..), and the cells will pick up property values from the style
unless they are locally set.