Add skeleton of two new classes to be used during table border painting:
TableBorderCollector, which is an abstract base class with a partial
implementation of a border collection strategy. And
CollapsingBorderCollector which is a concrete subclass for collecting
borders in the collapsing border model.

The idea here is that the table will have a collector configured, and
when painting itself it will

1) paint table background,
2) paint cell backgrounds,
3) use the border collector to collect the horizontal and vertical
   border into two lists,
4) if the table is configured to paint horizontal borders first:
     - iterate over horizontal borders and paint them,
     - iterate over vertical borders and paint them,
   and the other way around if vertical should be painter first.

There will be two collectors, one for use with the collapsed border
model and one for use with the separated border model, as these two
models are radically different. It's the former of these two that I've
added a skeleton for here now, as that's what I'll implement first.

This approach will

 - make it easy to flip around the order of border painting,
 - make it easy to specialize the behavior of how borders are
   constructed in the collapsed vs. the separated border model,
 - allow the collector to minimize the number of paints by merging
   subsequent borders if they are equal (in width/color).