Improve table resizing:
Added resize(...) function to PageItem_Table. The function takes a
width, height and a resize strategy that specifies how the resize should
affect the rows and columns. Currently implemented is the Equal strategy
which gives each row/column an equal amount of the change in
width/height, but Proportional is coming up. Rows and columns are never
resized below the new limits MinimumColumnWidth and MinimumRowHeight,
which are currently set to 3 pt.