Aesthetic improvements to border join cases 14 and 15:
When four borders meet at a cross and they can join pairwise, it looks
much better if the thicker of the two pairs win the conflict. So we
separate the two cases 14 and 15 into 14A, 14B, 15A and 15B. The image
on the wiki has been updated, as well as the Python test script that
exercises all (now 45) cases. I actually think we're doing a better job
than InDesign here now.