Fix painting of overflow marker and move it to PageItem:
The painting of the overflow marker was wrong. The size of it was
calculated by multiplying by the zoom factor, resulting in the marker
growing huge at large zoom levels. The code probably intended to
_divide_ by the zoom factor so that the on screen size of the marker
stays constant at different zoom levels. The behavior now is that at
zoom levels >= 100%, the size of the marker stays constant, and for zoom
levels < 100%, it shrinks with the zoom level. I also changed it to use
QPointF and the new drawLine(QPointF, QPointF) I've added to ScPainter.