Optimize painting of the table outline during resizing:
cellAt() and cellRect() are very fast, stroking is not. So paint the
outline in two passes, first left edges and then top edges. This allows
us to combine strokes where possible, which greatly reduces the number
of strokes, especially on homogenous tables, which are the norm.
Painting is now much faster and resizing tables with thousands of cells
no longer feels sluggish.