Use a dedicated initialize() function in table construction.
The function initializes the data structures with data for a first
row/column and only after that calls insertRows() and insertColumns()
to insert any remaining rows.

This allows for both insertRows() _and_ insertColumns() to assume that
rows() > 0 and columns() > 0, making them more symmetric, and avoids a
nasty infinite loop I was hitting while working on cell content (!).