Move PageItem_Table::cellRect to TableCell and rename it boundingRect():
Also add a stub for a contentRect() function. This will be implemented
shortly and will return the bounding rectangle of the cell adjusted for
paddings/borders.

From now on, lets use the terminology "bounding rectangle" for the
rectangle of the cell on the table grid, and "content rectangle" for
bounding rectangle adjusted for paddings/borders.