Initial work on cell content:
It's now possible to set text on a cell with setText(). The text
is represented by a PageItem_TextFrame for each cell. The text
will be correctly painted as part of normal table painting in
CollapsedTablePainter and borders/paddings are taken into account
when calculating the cell content rectangle.

If any borders/paddings on a cell/table is changed, or if a style
is set on a cell, or if the global cell/table style context in the
document changes, the cell text frames are updated to reflect the
change.

Other minor changes included in this commit are:

o Respect the View->Show Frames view setting and the frames color
  preference when painting table grid lines as a visual aid.