Handle activation of cells and typing of text into cells:
Cells are currently activated by clicking on them with the mouse.

The typing into cells is handled by passing on the event to the
handleModeEditKey(...) of the text frame of the currently active cell.

Cell selections are now temporarily disabled while I work on cursor
painting, text selection et.c. It will be brought back later.

NOTE: In order to get canvas updates I had to make CanvasMode_EditTable
      a friend of Canvas and set m_canvas->m_viewMode.forceRedraw = true
      before the call to m_canvas->update(). Will have to investigate if
      there's a better way.