Make sure text selection of cell text frames is painted:
The text frame would only draw its selection if in text edit mode, now
it will also paint it if in table edit mode. Another change was also
necessary to make it work; when a cell is activated with
setActiveCell(...), we call setSelected(true) on its text frame. This
will make the text frame think it is selected. I hope this will not have
any unwanted consequences, but I don't think it will.