Support setting the text cursor with the mouse in table cells:
ScribusView::slotSetCurs(...) had to be given support for tables; if the
current item is a table, the function will now use the text frame of the
currently active cell of the table, and the point to test at will be
mapped to table grid coordinates, since that's the coordinate system the
cell text frames are in.

Question; why is this function returning true on image frames? I've kept
this behavior, but I wonder why..