Fix crash when deleting a table and then quitting app.
Style sets would erronously emit the changed signal when deleted, and
since deleted items are destroyed after style sets, the table, which
listens to changes in the style sets, would be notified of this change
on application exit and try to access cell styles that had been deleted.

The fix is the one discussed on the mailing list; prevent the style set
from emitting the signal during destruction.