Fix crash / no initial cursor dilemma bug:
The PageItem::lastInFrame() function will return -1 when:

1) The frame is empty (no text).
2) The frame is so small that no text is visible (overflowing).

This led to a crash in CanvasMode::commonDrawTextCursor(...) which tried
to use this -1 as a parameter to StoryText::item(...). The old "fix"
that was in place for this in the edit canvas mode was to simply not
paint the cursor when lastInFrame() returns < 0. This fixed the crash
but had the unwanted side effect that no cursor was painted when
activating the edit mode on an empty frame.

This new fix instead changes CanvasMode::commonDrawTextCursor(...) to
still paint the cursor, but also make sure that it's never using a < 0
index.

(Note: CanvasMode::commonDrawTextCursor(...) is called
CanvasMode_Edit::drawTextCursor(...) in 1.5svn. It was moved/renamed in
my tables repository)