Fix selections/active cell when doing (semi-)destructive operations:
The behavior now is that when merging cells, the selection is cleared,
the new spanning cell is activated and we are returned to table edit
mode.

When removing rows or columns, any removed part of the selection is
cleared and the cell above or left of the removed rows or columns is
activated.

Also added some new asserts to assertValid() to make sure that we don't
end up with an invalid/covered active cell or invalid/covered selected
cells. Plus a check to make sure the active (logical) position is valid.